kumajala-backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
google-cloud-firestore==2.13.1
google-cloud-texttospeech==2.16.3
google-generativeai==0.3.2
python-dotenv==1.0.0
requests==2.31.0
Werkzeug==2.3.7